Fix for GAL-53, version resolution and single mvn repo manager
