+++ cli/src/main/java/org/jboss/galleon/cli/config/mvn/MavenMvnSettings.java
@@ -44,7 +44,7 @@ import org.eclipse.aether.RepositorySystemSession;
 import org.eclipse.aether.repository.RemoteRepository;
 import org.eclipse.aether.util.repository.AuthenticationBuilder;
 import org.jboss.galleon.ArtifactException;
-import static org.jboss.galleon.cli.MavenArtifactRepositoryManager.DEFAULT_REPOSITORY_TYPE;
+import static org.jboss.galleon.cli.CliMavenArtifactRepositoryManager.DEFAULT_REPOSITORY_TYPE;
 import org.jboss.galleon.cli.Util;
 
 /**
